Sprint 16: Code quality improvements and clippy warning fixes
- Fixed &PathBuf vs &Path parameter warnings
- Reduced clippy warnings from 58 to 42 (28% reduction)
- Updated function signatures for better performance
- Fixed type conversion issues
- Applied auto-fix suggestions

Following Toyota Way waste elimination (Muda) principles.

+++ server/src/handlers/tools.rs
@@ -1755,7 +1755,7 @@ async fn handle_analyze_defect_probability(
             analysis
                 .files
                 .into_iter()
-                .map(|f| (f.relative_path, f.churn_score as f32))
+                .map(|f| (f.relative_path, f.churn_score))
                 .collect()
         })
         .unwrap_or_default();
